Multiple changes...
Added connection ID to init plugin call. This precluded the need to make the default value of the `nextConnectionID` in `ServerHandler` start at -1 to denote a not-connected state. It then increments to 0 on first connect, retaining the original logic from then on.

Allow plugins to playback audio at any time. Moved around some definitions and code and changed logic on if audio buffer contains audio in the AudioOutput code.