FIX(channel-listener): Fix listener taking precedence over link
Before this commit it was possible to receive audio through a
ChannelListener (and thus with applied local volume adjustment) instead
of via the channel-link (and thus without the volume adjustment of a
listener proxy). This is fixed by this commit so that now the link
always takes precedence over the listener.