REFAC(client, talking-ui): Stylable widget wrapper
It turned out that it is absolutely impossible to only overwrite
specific properties via a styleSheet without also overwriting all other
properties that have been set via styleSheet on that object before.
Therefore a myWidget->setStyleSheet("background: red") will not only set
the background color to red but also erase everything that has been
written to the StyleSheet before this point.

As this is very inconvenient and always appending the most recent
property to the existing StyleSheet is not an option (the StyleSheet
will just keep growing in size and also this doesn't work well once one
starts using selectors), this commit introduces a wrapper class that can
be used for fine-tuned control of a widget's style.

This works by storing the desired properties in a separate class and
re-generate and re-apply the StyleSheet every time one of these
properties changes.

Currently the only implemented properties are font-size and
background-color.

NOTE: The wrapper only works as long as no other calls to setStyleSheet
are made to the object the wrapper acts upon.

The mentioned wrapper was then incorporated into the UI elements of the
TalkingUI.