FIX: Mic state coloring sticking to old channel
When changing from a channel in which someone was talking into a
different one, the coloring indicating the mic state of the local user
would still be applied to the old channel and as the local user was no
longer in it it also didn't get updated anymore.

Therefore we have to take some action if the local user changes his/her
channel in order to clear the coloring in the old channel and apply it
to the new one instead.