FEAT(client, talking-ui): Show whispers/shouts separately
Before w user that whispered or shouted to you was displayed to be
inside the channel they are currently in and the fact that they are
whispering/shouting was only visible by the color of their talking
icon.

In order to visually highlight whispers/shouts more, these are now
displayed in their own separate boxes in the TalkingUI. That means that
there can now be special "channels" (containers) that group all
currently whispering and all currently shouting users respectively.

These new containers are displayed before the regular channels in order
to easily see them.

As soon as a user stops whispering/shouting, they'll be moved back to
their regular channel box.