FEAT(client): Display listener icons in TalkingUI
Previously listeners of the local user were displayed as distinct
entries in the TalkingUI. This commit changes this and instead shows
channels a user has listeners in with an icon next to the channel name.
That icon can be selected and be used for triggering the usual context
menu.

The goal of this change is to save screen space for cases in which the
user may have many listeners placed.