FIX(client, talking-ui): Context-menu z-index issues
This issue was originally addressed by
c322a8e0f2fc8d7da8cb828bd15d17b8a03d11a6 but as it turned out the fix
was incomplete.

The issue of context menus being shown for a different entry than what
is selected in the TalkingUI persisted (at least for listeners) because
the position was stored in a separate variable as well (which got used
in the respective menu sub-functions) which was overlooked in the
original fix.