FEAT: Ability to query voice receiver (count)
With this protocol extension the client now has the possibility of
querying the server for how many users or which users are receiving the
audio if the user was to speak in a given channel or use one of its
defined shortcuts.

In principle this extension also allows for the server to push this
information to the client, should it decide to do so.