FEAT: Display audio receiver count in TalkingUI
If the server is new enough to support this feature, the TalkingUI will
display the amount of clients the local user is currently talking to.

This is meant to help identify situations in which one is accidentally
talking to a vastly different audience than anticipated (e.g.
accidentally shouting to the entire server) which allows for
reconsidering whether or not you want to continue speaking.

Additionally the user has the possibility of querying a list of users
that will be able to hear audio that is transmitted through the
currently active channel at the point the list was queried. That means
one can figure out exactly who is going to hear what one is about to
say.